Implement the api internals for circle layer based annotations. (#9)
* Implement the api internals for circle layer based annotations.

Implemented the base logic for source and layer creation in MGLAnnotationController.
When a controller is created, creates a source with an empty shape to be passed to the style layer. Each annotation’s controller child is responsible for creating the layer.

Implemented the basic internals for layout properties.

* Add circle annotation properties overrides.

Overrided the setters and getters for MGLCircleStyleAnnotation to store the values in an internal dictionary.

This dictionary will be passed as part of the Feature for styling.

* Add paint properties to MGLCircleStyleAnnotation.

Added circleStyle layer paint properties. Since there is a limit of the number of paint properties that can be used, the current implementation keeps track of each property that is set in the annotation. When a redraw is requested then the annotation’s controller enables the corresponding expression value.

* Add default values for circleLayer paint properties.

Added the default values according to the style specification to the MGLCircleStyleAnnotation properties.