Add annotation interaction support for tap events. (#16)
Added an UITapGestureRecognizer to hangle tap events. The current implementation reorganized the mapView recognizers, adding the newly created gesture to the array just befor the main mapView’s singleTapGestureRecognizer. The reason behind the reorg was due to an issue when selecting annotations. It was requiring two taps to select a new annotation when one was already selected.